source.py: fix leadtime and realization detection in EarthkitSource
We infer the number of realization from the duplicated leadtimes present
in every single chunk. Added several asserts to make sure we have
consistent detection. At the end of the day the only way to ensure
consistency in inferred realizations is requesting on a single-month
basis. The current implentation is very ad-hoc.

Signed-off-by: Jacopo Dall'Aglio <fingolfin00@gmail.com>